requirements.txt
# Core dependencies
hyperliquid-python-sdk>=0.14.1
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Async and scheduling
apscheduler==3.10.4
asyncio==3.4.3
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1